+++ src/Exception/IdTokenException.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types = 1);
 
-namespace HintikDev\OAuth2\Client\Exception;
+namespace Hintik\OAuth2\Client\Exception;
 
 use RuntimeException;